refactor authentication and user management with controller pattern and improved access control
