+++ backend/src/api/auth.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import { firebaseLogin, firebaseRegister } from "@/services/firebaseAuthService";
+import { AuthController } from "../controllers/AuthController";
 import { authSchema } from "@/validation/authSchema";
 
 const router = express.Router();
@@ -33,12 +33,8 @@ router.post("/login", async (req, res) => {
   if (!parseResult.success) {
     return res.status(400).json({ error: "Validation failed", details: parseResult.error.errors });
   }
-  const { email, password } = parseResult.data;
-  const result = await firebaseLogin(email, password);
-  if (result.success) {
-    return res.json(result.data);
-  }
-  return res.status(result.status ?? 500).json({ error: result.message, details: result.details });
+  req.body = parseResult.data;
+  return AuthController.login(req, res);
 });
 
 /**
@@ -70,12 +66,8 @@ router.post("/register", async (req, res) => {
   if (!parseResult.success) {
     return res.status(400).json({ error: "Validation failed", details: parseResult.error.errors });
   }
-  const { email, password } = parseResult.data;
-  const result = await firebaseRegister(email, password);
-  if (result.success) {
-    return res.json(result.data);
-  }
-  return res.status(result.status ?? 500).json({ error: result.message, details: result.details });
+  req.body = parseResult.data;
+  return AuthController.register(req, res);
 });
 
 /**
